modify tests to show behavior of 2 feature translation
the expectation was the translation would merge the
separate bridge and road features into a single osm
feature (since that is how it should be modeled in osm
schema).  this did not happen and the two features
remained distinct.  however the TDSv61 road feature
did get the bridge=yes tag, possibly due to its
combination of descriptive tags.